api: remove all C++ symbols from library => breaking C++ ABI change
There was a bug where sdb_find_by_identity would be exported by
glue/cplusplus.cpp, but the type might not match the header,
depending on the value of EB_STATUS_OR_VOID_T.

Fixing this requires making the method inline. However, that removes
the symbol from the library, which is a breaking change to the ABI
for C++. The API remains compatible, however.